+++ src/Entities/Games/CallbackGame.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * This file is part of the TelegramBot package.
+ *
+ * (c) Avtandil Kikabidze aka LONGMAN <akalongman@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Longman\TelegramBot\Entities\Games;
+
+use Longman\TelegramBot\Entities\Entity;
+
+/**
+ * Class CallbackGame
+ *
+ * A placeholder, currently holds no information. Use BotFather to set up your game.
+ *
+ * @link https://core.telegram.org/bots/api#callbackgame
+ **/
+class CallbackGame extends Entity
+{
+
+}